requirements.txt
Django==4.2
psycopg2-binary==2.9.6
gunicorn==20.1.0
django-rest-framework==0.1.0
django-extensions
celery==5.3.6
django-celery-beat
djangorestframework-simplejwt
